WFS-T support
This adds a generic FeatureServerClient to mercaido_client that should work with
any WFS-T server, and a GeoServerClient that integrates with the GeoServer REST
API to create and delete layers.

The configuration of Media, Recording and Feature Servers has been moved to a
separate page, and a page showing the layers currently present on a configured
Feature Server has been added.